+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- `--dump-args` CLI flag to render the other arguments in roughly the form expected by the library's `run` function.
+- Installation, usage, and versioning details to [README](./README.md).
 
 ## [0.3.2] - 2020-12-31
 ### Fixed